Enable rnd tools to fix lack of entropy
The AM335x has a hardware random number generator, but rngd needs to be
used to let it update the Linux kernel random number interfaces
(/dev/random and getrandom()). This is also needed to initialize
/dev/urandom in any reasonable time frame.

To use this, you need to start `rngd`. To compare, the nonblocking pool
is initialized roughly 40 seconds post boot with `rngd` and I didn't
wait long enough for it to initialize without it.