Add support for postponing firmware validation
This is not enabled by default. If you change the value of
nerves_fw_autovalidate to 1, then the Nerves application needs to set
nerves_fw_validated at some point before the first reboot. If it does
not, then the bootloader will revert to the previous software.

This is a major change to how the BBB boots and is incompatible with
previous builds. You can NOT run over the air firmware updates to
systems with this logic.