Compile everything with -O2
The Buildroot default is to optimize for size (-Os). This changes things
to pass -O2 to compile all C and C++ source code. Packages could have
overridden the optimization flags to use -O2 or -O3 already, but I've
only seen a few do this.  Changing this parameter results in ~500ms boot
time improvement (~3% with my current tests) on a different
single-processor ARM platform.