Add @github_organization
Whenever someone wants to create a custom package they need to modify the artifact and links section of the mix.exs. While it's trivial it's easy to forget and adds complications for new users.

With this change, making a custom image and artifact only requires informing them to change the @github_organization.