Switch timer to 100 Hz to match other systems
250 Hz seems to be a setting on Beagleboard kernel configurations, but
isn't used in the TI configurations. It's also not clear what benefit
using a non-default setting here has. Various internet searches turn up
either pre-NO_HZ articles or speculation on power savings.

This removes the non-default setting so that this system matches other
Nerves systems that use the defaults and the upstream TI systems that
use the defaults.