Move Linux kernel to SquashFS
The latest U-Boot's SquashFS support is much, much faster. This makes it
possible to remove the hack of moving the Linux kernel from the SquashFS
partition to the FATFS one. No more needing to write to the FATFS for
firmware updates!

FATFS read performance:

2828616 bytes read in 185 ms (14.6 MiB/s)

SquashFS read performance:

2828616 bytes read in 204 ms (13.2 MiB/s)

IMPORTANT: The main problem with this update is that if you have devices
deployed with U-Boot 2018.07 (everything with nerves_system_bbb v2.13.2
and earlier), then updating to this firmware works, but will be SLOW.