linux: revert to 5.10.168
This is a significant rollback of Linux kernel updates to avoid issues
that appear to have been introduced around or after 5.15. This is the
latest TI-patch release of the 5.10 series.

1. There's some issue that affects eMMC life time. See
   https://forum.beagleboard.org/t/possible-emmc-firmware-bug-or-hw-issue-recent-seeed-studio-bbbs-with-6-1-x-kernels/36648.
2. Wired Ethernet doesn't work on all Beaglebones and it appears to by
   related to the PHY driver. Searching the Beagleboard forum turns up a
   few possible causes.

Both issues seem to be tracked to changes in the Linux kernel after 5.10
(probably after 5.15, but this is currently less clear to me). The PHY
issue correlates to a TI CPSW change and the eMMC correlates to updates
to a MMC driver update. Both sound TI-specific.

Normally reverting Linux kernels wouldn't be done, but these issues are
very serious and it's unclear when fixes will be available.