Bump Linux to 6.12 LTS
It appears that the eMMC issue has been resolved so it's no longer
necessary to limit to 5.10. The official BeagleBone Debian images are on
6.12 as well.

The RT patches weren't available with this patchset. While RT support
was mainlined, it wasn't for 32-bit. If anyone actually uses RT support
with Nerves on BBB, it would be good to match their working
configuration. I don't think anyone actually does, though.